Rename ljumptab to ljumptabgcc because it uses GCC extensions
