+++ Pluto.vcxproj
@@ -194,7 +194,7 @@
     <ClInclude Include="src\ldo.h" />
     <ClInclude Include="src\lfunc.h" />
     <ClInclude Include="src\lgc.h" />
-    <ClInclude Include="src\ljumptab.h" />
+    <ClInclude Include="src\ljumptabgcc.h" />
     <ClInclude Include="src\llex.h" />
     <ClInclude Include="src\llimits.h" />
     <ClInclude Include="src\lmem.h" />
